requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.13
langgraph>=0.0.20
langchain-openai>=0.0.5

# Web Application
flask>=2.0.0
flask-session>=0.5.0
flask-cors>=3.0.10

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.4.0

# API Clients
openai>=1.1.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
tqdm>=4.65.0

# Development Tools
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0